# Add support for WAL preallocation