--- a/storage/syncing_file.go
+++ b/storage/syncing_file.go
@@ -8,16 +8,24 @@ import (
 	"sync/atomic"
 )
 
+// SyncingFileOptions holds the options for a syncingFile.
+type SyncingFileOptions struct {
+	BytesPerSync    int
+	PreallocateSize int
+}
+
 type syncingFile struct {
 	File
-	fd           int
-	useSyncRange bool
-	bytesPerSync int64
-	atomic       struct {
+	fd              uintptr
+	useSyncRange    bool
+	bytesPerSync    int64
+	preallocateSize int64
+	atomic          struct {
 		offset     int64
 		syncOffset int64
 	}
-	syncTo func(offset int64) error
+	preallocatedBlocks int64
+	syncTo             func(offset int64) error
 }
 
 // NewSyncingFile wraps a writable file and ensures that data is synced
@@ -26,21 +34,28 @@ type syncingFile struct {
 // decides to write out a large chunk of dirty filesystem buffers. If
 // bytesPerSync is zero, the original file is returned as no syncing is
 // requested.
-func NewSyncingFile(f File, bytesPerSync int) File {
-	if bytesPerSync <= 0 {
-		return f
-	}
+func NewSyncingFile(f File, opts SyncingFileOptions) File {
 	s := &syncingFile{
-		File:         f,
-		fd:           -1,
-		bytesPerSync: int64(bytesPerSync),
+		File:            f,
+		bytesPerSync:    int64(opts.BytesPerSync),
+		preallocateSize: int64(opts.PreallocateSize),
 	}
+
+	type fd interface {
+		Fd() uintptr
+	}
+	if d, ok := f.(fd); ok {
+		s.fd = d.Fd()
+	}
+
 	s.init()
 	return s
 }
 
 // NB: syncingFile.Write is unsafe for concurrent use!
 func (f *syncingFile) Write(p []byte) (n int, err error) {
+	_ = f.preallocate(atomic.LoadInt64(&f.atomic.offset) + int64(n))
+
 	n, err = f.File.Write(p)
 	if err != nil {
 		return n, err
@@ -52,6 +67,22 @@ func (f *syncingFile) Write(p []byte) (n int, err error) {
 		return 0, err
 	}
 	return n, nil
+}
+
+func (f *syncingFile) preallocate(offset int64) error {
+	if f.fd == 0 || f.preallocateSize == 0 {
+		return nil
+	}
+
+	newPreallocatedBlocks := (offset + f.preallocateSize - 1) / f.preallocateSize
+	if newPreallocatedBlocks <= f.preallocatedBlocks {
+		return nil
+	}
+
+	length := f.preallocateSize * (newPreallocatedBlocks - f.preallocatedBlocks)
+	offset = f.preallocateSize * f.preallocatedBlocks
+	f.preallocatedBlocks = newPreallocatedBlocks
+	return preallocExtend(f.fd, offset, length)
 }
 
 func (f *syncingFile) ratchetSyncOffset(offset int64) {
@@ -76,6 +107,10 @@ func (f *syncingFile) Sync() error {
 }
 
 func (f *syncingFile) maybeSync() error {
+	if f.bytesPerSync <= 0 {
+		return nil
+	}
+
 	// From the RocksDB source:
 	//
 	//   We try to avoid sync to the last 1MB of data. For two reasons:
@@ -98,7 +133,7 @@ func (f *syncingFile) maybeSync() error {
 		return nil
 	}
 
-	if f.fd < 0 {
+	if f.fd == 0 {
 		return f.Sync()
 	}
 	return f.syncTo(syncToOffset)