On Linux, preallocation makes a huge difference in sync performance.

```
name                         time/op
SyncPreallocFile/size=0MB-4  503µs ± 6%
SyncPreallocFile/size=4MB-4  303µs ± 2%
SyncReuseFile-4              291µs ± 1%
```

The effect is much more mutated on Darwin, but still present:

```
name                         time/op
SyncPreallocFile/size=0MB-8  50.4µs ±14%
SyncPreallocFile/size=4MB-8  46.4µs ± 2%
SyncReuseFile-8              46.7µs ± 2%
```

Note that the `SyncReuseFile` benchmark is an attempt to measure the win from reusing WAL files. WAL file reuse is not actually implemented  yet.

See #41