@petermattis 2019-04-15
Looks like I missed up the perf numbers on Linux as I was measuring `/tmp`. The updated numbers show that preallocation and reuse are fairly close in speed. Here is a more direct view of the delta:

```
name        old time/op  new time/op  delta
SyncFile-4   312µs ± 3%   299µs ± 4%  -4.09%  (p=0.000 n=18+20)
```

"Old" is with preallocation. "New" is with reusing the same file over and over (and no preallocation).

I've verified that I'm definitely using `fdatasync` in both cases.