@vtolstov 2019-04-15
Does it possible to unmap from disk some of unneeded wal data? (mostly question for ssd case)