@petermattis 2019-04-15
@vtolstov WAL files are currently deleted when they are no longer needed. I'm not sure what you're asking. Can you rephrase your question?