@vtolstov 2019-04-15
ah sorry, in this case unmap can happen then file is removed from fs and fs driver code send unmap to this area, sorry.