@ajkr 2019-04-17
@vtolstov I think if the filesystem is mounted with discard option, then the unmap/trim command should be sent to the device when the file is deleted. Let me know if I misunderstood your question.